Mengubah class Employee
Mengurangi jumlah parameter pada Employee
