+++ Staff.java
@@ -1,0 +1,12 @@
+package lib;
+
+public class Staff extends TahunGabung {
+	
+	private String employeeId;
+	private String firstName;
+	private String lastName;
+	private String idNumber;
+	private String address;
+	
+	
+}
