+++ TahunGabung.java
@@ -1,0 +1,10 @@
+package lib;
+
+public class TahunGabung extends GajiPerbulan {
+	
+	private int yearJoined;
+	private int monthJoined;
+	private int dayJoined;
+	private int monthWorkingInYear;
+	
+}
